refactor(service): make `Cgroup` like other C-style enums
